% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_fitdist.R
\name{calc_bic}
\alias{calc_bic}
\title{Calculates the Bayesian information criterion for the loglikelihood of a two
parameter probability distribution}
\usage{
calc_bic(loglik, data, df = 2)
}
\arguments{
\item{loglik}{A vector or single number the loglikelihood of the model}

\item{data}{A numeric vector of values}

\item{df}{A numeric specifying the degrees of freedom for the model in order
to calculate the Akaike information criterion}
}
\value{
A single or vector of numerics equal to the input vector length
}
\description{
Calculates the Bayesian information criterion for the loglikelihood of a two
parameter probability distribution
}
\examples{
# example using vector
data <- c(2, 13, 22, 25, 11, 12, 11, 23, 13, 24)
calc_bic(loglik = -110, data = data)

# example using tabular data
data <- data.frame(
  c(2, 13, 22, 25, 11, 12, 11, 23, 13, 24),
  c(4, 15, 19, 1, 16, 10, 3, 17, 16, 3)
)
calc_bic(loglik = -110, data = data)
}
