% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_profiles.R
\name{calculate_profile}
\alias{calculate_profile}
\title{Estimate profile likelihood and 95% CI}
\usage{
calculate_profile(
  log_likelihood,
  data_in,
  n_param,
  a_initial,
  b_initial,
  precision = 0.1
)
}
\arguments{
\item{log_likelihood}{Log-likelihood function in form \code{function(x,a)} (one parameter model) or \code{function(x,a,b)} (two parameter model)}

\item{data_in}{Vector of observations to be evaluated in log_likelihood, with overall likelihood given by sum(log_likelihood)}

\item{n_param}{Number of parameters in \code{log_likelihood} model}

\item{precision}{Parameter defining how fine-scale the grid search is for the profile likelihood}

\item{a_inital}{Initial guess for parameter \code{a}}

\item{b_inital}{Initial guess for parameter \code{b} (if a two parameter model, otherwise default is NULL)}
}
\description{
This function estimate the profile likelihood and 95% CI for a one or two parameter model
}
\examples{
calculate_profile()
}
