% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checkers.R
\name{function_checkers}
\alias{function_checkers}
\alias{test_func_req_args}
\alias{test_func_num_out}
\title{Check functions passed to exported functions}
\usage{
test_func_req_args(func, n_req_args = 1)

test_func_num_out(func, n = 10)
}
\arguments{
\item{func}{A function. This is expected to be a function evaluating the
density of a distribution at numeric values.}

\item{n_req_args}{The number of required arguments, i.e., arguments without
default values.}

\item{n}{The number of elements over which to evaluate the function \code{func}.
Defaults to 10, and \code{func} is evaluated over \code{seq(n)}.}
}
\value{
A logical for whether the function \code{func} fulfils conditions
specified in the respective checks.
}
\description{
Internal helper function that check whether a function passed as
an argument meets the requirements of package methods.

\code{test_fn_req_args()} checks whether the function has only the expected number
of required arguments, i.e., arguments without default values. Defaults to
checking for a single required argument.

\code{test_fn_num_out()} checks whether a function accepting a numeric vector
input returns a numeric vector output of the same length as the input, with
finite non-missing values \eqn{\geq} 0.0.
}
\keyword{internal}
