% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_profile.R
\name{calculate_profile}
\alias{calculate_profile}
\title{Estimate profile likelihood and 95\% CI}
\usage{
calculate_profile(
  log_likelihood,
  data_in,
  n_param,
  a_initial,
  b_initial = NULL,
  precision = 0.01
)
}
\arguments{
\item{log_likelihood}{Log-likelihood function in form \verb{function(x,a)}
(one parameter model) or \verb{function(x,a,b)} (two parameter model)}

\item{data_in}{Vector of observations to be evaluated in log_likelihood,
with overall likelihood given by sum(log_likelihood)}

\item{n_param}{Number of parameters in \code{log_likelihood} model}

\item{a_initial}{Initial guess for parameter \code{a}}

\item{b_initial}{Initial guess for parameter \code{b} (if a two parameter
model, otherwise default is NULL)}

\item{precision}{Parameter defining how fine-scale the grid search is for
the profile likelihood}
}
\description{
This function estimate the profile likelihood and 95\% CI for a
one or two parameter model
}
\examples{
# example of a one parameter model
data_in <- rnorm(n = 100, mean = 1, sd = 2)
log_likelihood <- function(x,a) dnorm(x, a, 2, log = TRUE)
n_param <- 1
a_initial <- 0.5

calculate_profile(
  log_likelihood = log_likelihood,
  data_in = data_in,
  n_param = 1,
  a_initial = 4,
  precision = 0.1
)

# examle of a two parameter model
data_in <- rnorm(n = 100, mean = 1, sd = 1)
log_likelihood <- function(x,a,b) dnorm(x, a, b, log = TRUE)
n_param <- 2
a_initial <- 0.5
b_initial <- 1

calculate_profile(
  log_likelihood = log_likelihood,
  data_in = data_in,
  n_param = 2,
  a_initial = 4,
  b_initial = 2,
  precision = 0.1
)
}
\author{
Adam Kucharski, Joshua W. Lambert
}
