% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mle.R
\name{estimate_mle}
\alias{estimate_mle}
\title{Calculates the maximum-likelihood estimate for a given likelihood function}
\usage{
estimate_mle(log_likelihood, data_in, n_param, a_initial, b_initial = NULL)
}
\arguments{
\item{log_likelihood}{Log-likelihood function in form \verb{function(x, a)}
(one parameter model) or \verb{function(x, a, b)} (two parameter model)}

\item{data_in}{Vector of observations to be evaluated in log_likelihood,
with overall likelihood given by sum(log_likelihood)}

\item{n_param}{Number of parameters in \code{log_likelihood} model}

\item{a_initial}{Initial guess for parameter \code{a}}

\item{b_initial}{Initial guess for parameter \code{b} (if a two parameter
model, otherwise default is NULL)}
}
\description{
This function calculates the maximum-likelihood estimate for a
one or two parameter model
}
\examples{
# example for a one parameter model
sim_data <- rnorm(100, 5, 2)
log_likelihood <- function(x, a) dnorm(x, a, 2, log = TRUE)
a_initial <- 4

estimate_mle(
  log_likelihood = log_likelihood,
  data_in = sim_data,
  n_param = 1,
  a_initial = 1
)

# example for a two parameter model
sim_data <- rnorm(100, 5, 2)
log_likelihood <- function(x, a, b) dnorm(x, a, b, log = TRUE)
a_initial <- 4
b_initial <- 1

estimate_mle(
  log_likelihood = log_likelihood,
  data_in = sim_data,
  n_param = 2,
  a_initial = 1,
  b_initial = 1
)
}
\author{
Adam Kucharski, Joshua W. Lambert
}
